myDjangoProject/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
djangorestframework==3.14.0
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
numpy==1.23.5
pip==22.3.1
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
setuptools==63.2.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
